@@ -3,9 +3,9 @@
   "version": "0.0.1",
   "description": "Test suite to demonstrate interoperability between various VC issuers & verifiers.",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "parallel": "npm run generate-vdls && mocha tests/ --require esm --reporter @digitalbazaar/mocha-w3c-interop-reporter --reporter-options body=\"$PWD/body.hbs\",matrix=\"$PWD/matrix.hbs\",reportDir=\"$PWD/reports\",respec=\"$PWD/respecConfig.json\",title=\"Verifiable Driver's License Interoperability Report 1.0\",helpers=\"$PWD/templateHelpers.js\",suiteLog='./suite.log' --parallel --timeout 15000",
-    "test": "npm run generate-vdls && mocha tests/ --require esm --reporter @digitalbazaar/mocha-w3c-interop-reporter --reporter-options body=\"$PWD/body.hbs\",matrix=\"$PWD/matrix.hbs\",reportDir=\"$PWD/reports\",respec=\"$PWD/respecConfig.json\",title=\"Verifiable Driver's License Interoperability Report 1.0\",helpers=\"$PWD/templateHelpers.js\",suiteLog='./suite.log' --timeout 15000",
+    "test": "mocha tests/ --reporter @digitalbazaar/mocha-w3c-interop-reporter --reporter-options body=\"$PWD/body.hbs\",matrix=\"$PWD/matrix.hbs\",reportDir=\"$PWD/reports\",respec=\"$PWD/respecConfig.json\",title=\"Verifiable Driver's License Interoperability Report 1.0\",helpers=\"$PWD/templateHelpers.cjs\",suiteLog='./suite.log' --timeout 15000",
     "generate-vdls": "node generate-vdls.js",
     "lint": "eslint *.js",
     "postinstall": "cd bbs && npm i"
@@ -30,28 +30,29 @@
   },
   "homepage": "https://github.com/w3c-ccg/vdl-test-suite#readme",
   "dependencies": {
-    "@digitalbazaar/did-method-key": "^1.0.0",
-    "@digitalbazaar/ed25519-signature-2020": "^2.1.0",
-    "@digitalbazaar/ed25519-verification-key-2020": "^2.1.1",
-    "@digitalbazaar/http-client": "^1.0.0",
-    "@digitalbazaar/mocha-w3c-interop-reporter": "^1.0.0",
+    "@digitalbazaar/did-method-key": "^3.0.0",
+    "@digitalbazaar/ed25519-signature-2020": "^5.0.0",
+    "@digitalbazaar/ed25519-verification-key-2020": "^4.1.0",
+    "@digitalbazaar/http-client": "^3.2.0",
+    "@digitalbazaar/mocha-w3c-interop-reporter": "^1.3.0",
     "@digitalbazaar/security-context": "^1.0.0",
-    "@digitalbazaar/vc": "^1.0.0",
-    "@digitalbazaar/vpqr": "^2.1.0",
+    "@digitalbazaar/vc": "^5.0.0",
+    "@digitalbazaar/vpqr": "^3.0.0",
     "axios": "^0.21.1",
     "chai": "^4.3.4",
     "cit-context": "^2.0.0",
-    "credentials-context": "^1.0.0",
+    "credentials-context": "^2.0.0",
     "csv-parse": "^4.15.3",
     "did-context": "^3.0.1",
     "ed25519-signature-2020-context": "^1.0.1",
-    "esm": "^3.2.25",
     "file-size": "^1.0.0",
     "jsonld-document-loader": "^1.1.0",
+    "klona": "^2.0.5",
     "mocha": "^8.3.2",
     "node-cbor": "^5.0.4",
     "require-dir": "^1.2.0",
     "uuid": "^8.3.2",
+    "vc-api-test-suite-implementations": "github:w3c-ccg/vc-api-test-suite-implementations",
     "vc-revocation-list-context": "^1.0.0",
     "vdl-context": "github:digitalbazaar/vdl-context",
     "x25519-key-agreement-2020-context": "^1.0.0"
@@ -60,8 +61,8 @@
     "node": ">=14.0.0"
   },
   "devDependencies": {
-    "eslint": "^7.23.0",
-    "eslint-config-digitalbazaar": "^2.6.1",
-    "eslint-plugin-jsdoc": "^32.3.0"
+    "eslint": "^8.27.0",
+    "eslint-config-digitalbazaar": "^4.1.0",
+    "eslint-plugin-jsdoc": "^39.6.2"
   }
 }
