@@ -3,12 +3,14 @@
  */
 'use strict';
 
-const {join} = require('path');
-const {writeJSON} = require('./files');
-const {paths} = require('./paths');
-const {CONTEXT_URL} = require('vdl-context');
-const stateList = require('./states');
-const didKeyDriver = require('@digitalbazaar/did-method-key').driver();
+import {CONTEXT_URL} from 'vdl-context';
+import {driver} from '@digitalbazaar/did-method-key';
+import {join} from 'path';
+import {paths} from './paths.js';
+import stateList from './states.json' assert {type: 'json'};
+import {writeJSON} from './files.js';
+
+const didKeyDriver = driver();
 
 async function createVC(state) {
   const contexts = [
