@@ -17,14 +17,14 @@ import {
   BbsBlsSignatureProof2020,
   deriveProof
 } from '@mattrglobal/jsonld-signatures-bbs';
-import {extendContextLoader, sign, verify, purposes} from 'jsonld-signatures';
+import jsigs from 'jsonld-signatures';
 
-import _keyPairOptions from './data/keyPair.json';
-import _disclosures from './data/deriveProofFrame.json';
-import exampleControllerDoc from './data/controllerDocument.json';
-import bbsContext from './data/bbs.json';
-import credentialContext from './data/credentialsContext.json';
-import jwsContext from './data/jwsContext.json';
+import _keyPairOptions from './data/keyPair.json' assert {type: 'json'};
+import _disclosures from './data/deriveProofFrame.json' assert {type: 'json'};
+import exampleControllerDoc from './data/controllerDocument.json' assert {type: 'json'};
+import bbsContext from './data/bbs.json' assert {type: 'json'};
+import credentialContext from './data/credentialsContext.json' assert {type: 'json'};
+import jwsContext from './data/jwsContext.json' assert {type: 'json'};
 import {CONTEXT_URL as vdlContextUri, CONTEXT} from 'vdl-context';
 
 const documents = {
@@ -56,7 +56,7 @@ const customDocLoader = url => {
 };
 
 //Extended document load that uses local contexts
-const _documentLoader = extendContextLoader(customDocLoader);
+const _documentLoader = jsigs.extendContextLoader(customDocLoader);
 
 /**
  * Creates a BBS+ report for a VC.
@@ -81,16 +81,16 @@ export const createBBSreport = async ({
   const keyPair = await new Bls12381G2KeyPair(keyPairOptions);
 
   //Sign the input document
-  const signedDocument = await sign(inputDocument, {
+  const signedDocument = await jsigs.sign(inputDocument, {
     suite: new BbsBlsSignature2020({key: keyPair}),
-    purpose: new purposes.AssertionProofPurpose(),
+    purpose: new jsigs.purposes.AssertionProofPurpose(),
     documentLoader
   });
 
   //Verify the proof
-  let verified = await verify(signedDocument, {
+  let verified = await jsigs.verify(signedDocument, {
     suite: new BbsBlsSignature2020(),
-    purpose: new purposes.AssertionProofPurpose(),
+    purpose: new jsigs.purposes.AssertionProofPurpose(),
     documentLoader
   });
   //Derive a proof
@@ -100,9 +100,9 @@ export const createBBSreport = async ({
   });
 
   //Verify the derived proof
-  verified = await verify(derivedProof, {
+  verified = await jsigs.verify(derivedProof, {
     suite: new BbsBlsSignatureProof2020(),
-    purpose: new purposes.AssertionProofPurpose(),
+    purpose: new jsigs.purposes.AssertionProofPurpose(),
     documentLoader
   });
   return {
